// src/app/enums/enum-genres.ts
export enum Genres {
    Action = 28,
    Adventure = 12,
    Animation = 16,
    Comedy = 35,
    Crime = 80,
    Horror = 27,
    Drama = 18,
    Documentary = 99,
    Family = 10751,
    Fantasy = 14,
    Mystery = 9648,
    History = 36,
    Romance = 10749,
    ScienceFiction = 878,
    Music = 10402,
    TVMovie = 10770,
    Thriller = 53,
    War = 10752,
    Western = 37
}